requirements.txt
Flask
Flask-Session
werkzeug.security
Flask-SQLAlchemy
flask-login
os
flask_wtf
email_validator
bootstrap-icons 